+++ src/Caster/Caster.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Caster;
+
+interface Caster
+{
+    public static function castToDatabase($value): ?string;
+
+    public function castFromDatabase(?string $value);
+}