+++ tests/Functional/AbstractFunctionalTestCase.php
@@ -1,0 +1,29 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Tests\Functional;
+
+use Php\Support\Laravel\Tests\AbstractTestCase;
+
+/**
+ * Class AbstractFunctionalTestCase
+ */
+abstract class AbstractFunctionalTestCase extends AbstractTestCase
+{
+    /**
+     * @return void
+     */
+    protected function setUp(): void
+    {
+        parent::setUp();
+    }
+
+    /**
+     * @return void
+     */
+    protected function tearDown(): void
+    {
+        $this->artisan('db:wipe');
+    }
+}