+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+## [0.1.0] - 2020-01-29
+### Changed
+- Refactor `Caster` functionality.
+- Remove from trait `PostgresArray` methods: `mutateToPgArray`, `accessPgArray`.
+
+### Added
+- Written tests.
+
+### Removed
+- Trait `CasterAttribute`.
+
+
 ## [0.0.3] - 2020-01-28
 ### Changed
 - Move Trait `CasterAttribute` to different namespace: `Php\Support\Laravel\Traits\Models`.
@@ -25,7 +37,8 @@ and this project adheres to [Semantic Versioning][semver].
 - Trait for DB/Migration `UUID`. 
 
 [Unreleased]: https://github.com/efureev/laravel-support/compare/v0.0.2...HEAD
-[0.0.2]: https://github.com/efureev/laravel-support/compare/v0.0.2...v0.0.3
+[0.1.0]: https://github.com/efureev/laravel-support/compare/v0.0.3...v0.1.0
+[0.0.3]: https://github.com/efureev/laravel-support/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/efureev/laravel-support/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/efureev/laravel-support/releases/tag/v0.0.1